+++ src/app/ApplicationTemplate.Access/NativePlatform/Connectivity/ConnectivityChangedEventArgs.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace ApplicationTemplate.DataAccess;
+
+public sealed class ConnectivityChangedEventArgs : EventArgs
+{
+	public ConnectivityChangedEventArgs(ConnectivityState state)
+	{
+		State = state;
+	}
+
+	public ConnectivityState State { get; }
+}