+++ src/app/ApplicationTemplate.Access/NativePlatform/Email/EmailBodyFormat.cs
@@ -1,4 +1,4 @@
-﻿namespace ApplicationTemplate;
+﻿namespace ApplicationTemplate.DataAccess;
 
 public enum EmailBodyFormat
 {