+++ src/app/ApplicationTemplate.Shared.Views/Configuration/ApiConfiguration.cs
@@ -1,5 +1,6 @@
 ﻿using System.Net.Http;
 using System.Threading.Tasks;
+using ApplicationTemplate.DataAccess;
 using MallardMessageHandlers;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -13,8 +14,8 @@ public static class ApiConfiguration
 			.AddMainHandler()
 			.AddSingleton<INetworkAvailabilityChecker>(serviceProvider =>
 			{
-				var connectivityProvider = serviceProvider.GetRequiredService<IConnectivityProvider>();
-				return new NetworkAvailabilityChecker(_ => Task.FromResult(connectivityProvider.NetworkAccess is NetworkAccess.Internet));
+				var connectivityProvider = serviceProvider.GetRequiredService<IConnectivityRepository>();
+				return new NetworkAvailabilityChecker(_ => Task.FromResult(connectivityProvider.State is ConnectivityState.Internet));
 			});
 	}