+++ src/app/ApplicationTemplate.Shared.Views/Configuration/ViewServicesConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reactive.Concurrency;
+using ApplicationTemplate.DataAccess;
 using Chinook.DynamicMvvm;
 using MessageDialogService;
 using Microsoft.Extensions.DependencyInjection;
@@ -30,8 +31,8 @@ public static class ViewServicesConfiguration
 			.AddSingleton<IAppStoreUriProvider, AppStoreUriProvider>()
 			.AddSingleton<IDeviceInformationProvider, DeviceInformationProvider>()
 			.AddSingleton<IExtendedSplashscreenController, ExtendedSplashscreenController>(s => new ExtendedSplashscreenController(Shell.Instance.DispatcherQueue))
-			.AddSingleton<IConnectivityProvider, ConnectivityProvider>()
-			.AddSingleton<IEmailService, EmailService>()
+			.AddSingleton<IConnectivityRepository, ConnectivityRepository>()
+			.AddSingleton<IEmailRepository, EmailRepository>()
 			.AddSingleton<IMemoryProvider, MemoryProvider>()
 			.AddSingleton<IReviewPrompter, ReviewPrompter>()
 			.AddMessageDialog();